Separate model in model.go into sub file
